ai/requirements.txt
# ─────────────────────────────────────────────────────────────
# Core serving stack
# ─────────────────────────────────────────────────────────────
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ─────────────────────────────────────────────────────────────
# Deep-learning & vision (CPU-only)
# ─────────────────────────────────────────────────────────────
torch==2.2.0
torchvision==0.17.0
pillow==10.3.0
numpy==1.26.4

# ─────────────────────────────────────────────────────────────
# Training-time utilities
# ─────────────────────────────────────────────────────────────
scikit-learn==1.4.2
matplotlib==3.8.4
tqdm==4.66.2
torchmetrics>=1.3
optuna>=3.6

# ─────────────────────────────────────────────────────────────
# Supabase SDK
# ─────────────────────────────────────────────────────────────
supabase==2.15.1
storage3==0.11.3

# ─────────────────────────────────────────────────────────────
# Misc
# ─────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ─────────────────────────────────────────────────────────────
# Model code
# ─────────────────────────────────────────────────────────────
ultralytics==8.3.142
